HailTaskContext.get => SparkTaskContext.get
